requirements.txt
streamlit>=1.41.0
Pillow>=10.0.0
streamlit-cropper>=0.2.1
pikepdf>=8.0.0
pypdf>=3.17.0